user can manually specify species_order, whether to transform, and move progbar to utils
